refactor: Change minutes input EditText to a TextView.
Countdown time limit is between 1 to 3 minutes, inclusive.

To make more consistent with rest of the PromptActivity screen.
The prompt text is automatically chosen randomly and so is the image,
so the minutes can also be randomly chosen too.

Ran on API 28 emulator and API 17 tablet to test manually.
Rotated screen to check minutes text was the same.