fix: Timer resumes after user leaves app and returns to it.
When user leaves app, ex. by pressing home button,
timer was not starting again.
Timer was being canceled in onStop, but onCreate() was not called if
if user went to home screen, so no new timer was created or started.

Solution was starting timer in onStart() instead.

Removed onRestoreInstanceState() and KEY_END_TIME since not needed.

App immediately resumes timer if press back in OutOfTimeActivity or CompletedOnTimeActivity.

Runs on API 17 tablet.