feat: Add initial RecyclerView to MyWritingFragment.
Renamed rv_writing to rv_my_writing and writing_row to my_writing_row.
To avoid confusing these components with WritingFragment.

Created adapter and initialized a default list of items in MyWritingFragment.

Added item divider.

Set empty text to not be visible if there are items in the list.

Runs on API 28 emulator and API 17 tablet. UI tests pass on API 17 tablet.