fix: Listen for changes after user signs out and signs back in.
Firestore was giving permission denied snapshot error after signing out and
was not listening for recieved or sent challenges, fix by setting boolean isListeningForChallenges
and isListenFunctionCalled to false when there's snapshot error and calling function again in
HomeFragment.

Manually tested on Android 4.2.2 tablet and 7.0 phone with existing user accounts which have writings and challenges.
Tested display notifications of receive and completed challenge when already signed in, and
after signing out and signing in without quitting.